move func to secrets record
