+++ checkov/common/output/secrets_record.py
@@ -1,5 +1,11 @@
 from __future__ import annotations
+
+import logging
 from typing import Optional, List, Tuple, Dict, Any
+
+from termcolor import colored
+
+from checkov.secrets.consts import ValidationStatus
 
 from checkov.common.bridgecrew.severities import Severity
 
@@ -63,3 +69,21 @@ class SecretsRecord(Record):
                          definition_context_file_path=definition_context_file_path
                          )
         self.validation_status = validation_status
+
+    def _get_secret_validation_status_message(self) -> str:
+        warning_sign_unicode = '\u26a0'
+        text_by_secret_validation_status_status = {
+            ValidationStatus.VALID.value: colored(f'\t{warning_sign_unicode} This secret has been validated'
+                                                  f' and should be prioritized\n', "red"),
+            ValidationStatus.INVALID.value: '\tThis is not a valid secret and can be de-prioritized\n',
+            ValidationStatus.UNKNOWN.value: '\tWe were not able to validate this secret\n',
+            ValidationStatus.UNAVAILABLE.value: ''
+        }
+        message = None
+        if hasattr(self, 'validation_status'):
+            message = text_by_secret_validation_status_status.get(self.validation_status)
+
+            if not message and self.validation_status != ValidationStatus.UNAVAILABLE.value:
+                logging.debug(f'Got empty message for secret validation status = {self.validation_status}')
+
+        return message or ''
